+++ packages/cedar-amcharts/test/render/render.spec.ts
@@ -1,8 +1,10 @@
 import 'amcharts3/amcharts/amcharts'
 import 'amcharts3/amcharts/serial'
-import {} from 'jest'
+import { } from 'jest'
 import render from '../../src/render/render'
 import bar from '../../src/specs/bar'
+import pie from '../../src/specs/pie'
+import scatter from '../../src/specs/scatter'
 import timeline from '../../src/specs/timeline'
 import barSpec from '../data/barSpec'
 import builtBarSpec from '../data/builtBarSpec'
@@ -18,16 +20,268 @@ describe('Spec gets fetched properly', () => {
   })
 })
 
-describe('Spec gets filled in properly', () => {
-  test('A simple spec gets filled in properly', () => {
+describe('when filling in a bar spec', () => {
+  test('it should get filled in properly', () => {
     const spec = (bar as any)
     spec.dataProvider = barSpec.realData
-    spec.categoryField = 'categoryField'
     const result = render.fillInSpec(spec, barSpec.spec)
-    expect(result.theme).toBe('calcite')
     expect(result).toEqual(builtBarSpec)
   })
 })
+
+describe('when filling in a scatter spec', () => {
+  let result
+  beforeAll(() => {
+    const spec = (scatter as any)
+    /* tslint:disable */
+    const definition = {
+      "type": "scatter",
+      "datasets": [ {
+        "url": "https://services1.arcgis.com/bqfNVPUK3HOnCFmA/arcgis/rest/services/Demographics_(Median_Household_Income)/FeatureServer/0"
+      }],
+      "series": [
+        {
+          "category": {"field": "TotalPop2015", "label": "Population"},
+          "value": {"field": "MedianHHIncome2015", "label": "Median Median Household Income"}
+        }
+      ]
+    }    
+    /* tslint:enable */
+    result = render.fillInSpec(spec, definition)
+  })
+  test('it should set the graph xField field', () => {
+    expect(result.graphs[0].xField).toEqual('TotalPop2015')
+  })
+  test('it should set the graph yField field', () => {
+    expect(result.graphs[0].yField).toEqual('MedianHHIncome2015')
+  })
+  test('should set x axis title', () => {
+    const xAxis = result.valueAxes[0]
+    expect(xAxis.title).toEqual('Population')
+    expect(xAxis.position).toEqual('bottom')
+  })
+  test('should set y axis title', () => {
+    const yAxis = result.valueAxes[1]
+    expect(yAxis.title).toEqual('Median Median Household Income')
+    expect(yAxis.position).toEqual('left')
+  })
+})
+
+describe('when filling in a stacked bar spec', () => {
+  let result
+  beforeAll(() => {
+    const spec = (bar as any)
+    /* tslint:disable */
+    const definition = {
+      "type": "bar",
+      "datasets": [
+        {
+          "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
+          "name": "Jordan",
+          "query": {
+            "where": "City='Jordan'",
+            "orderByFields": "Number_of_SUM DESC",
+            "groupByFieldsForStatistics": "Type",
+            "outStatistics": [{
+              "statisticType": "sum",
+              "onStatisticField": "Number_of",
+              "outStatisticFieldName": "Number_of_SUM"
+            }]
+          },
+          "join": "Type"
+        },
+        {
+          "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
+          "name": "Dewitt",
+          "query": {
+            "where": "City='Dewitt'",
+            "orderByFields": "Number_of_SUM DESC",
+            "groupByFieldsForStatistics": "Type",
+            "outStatistics": [{
+              "statisticType": "sum",
+              "onStatisticField": "Number_of",
+              "outStatisticFieldName": "Number_of_SUM"
+            }]
+          },
+          "join": "Type"
+        },
+        {
+          "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
+          "name": "Fayetteville",
+          "query": {
+            "where": "City='Fayetteville'",
+            "orderByFields": "Number_of_SUM DESC",
+            "groupByFieldsForStatistics": "Type",
+            "outStatistics": [{
+              "statisticType": "sum",
+              "onStatisticField": "Number_of",
+              "outStatisticFieldName": "Number_of_SUM"
+            }]
+          },
+          "join": "Type"
+        }
+      ],
+      "series": [
+        {
+          "category": {"field": "Type", "label": "Type"},
+          "value": { "field": "Number_of_SUM", "label": "Jordan Students"},
+          "source": "Jordan",
+          "stack": true
+        },
+        {
+          "category": {"field": "Type", "label": "Type"},
+          "value": { "field": "Number_of_SUM", "label": "Dewitt Students"},
+          "source": "Dewitt",
+          "stack": true
+        },
+        {
+          "category": {"field": "Type", "label": "Type"},
+          "value": { "field": "Number_of_SUM", "label": "Fayetteville Students"},
+          "source": "Fayetteville",
+          "stack": true
+        }
+      ]
+    }    
+    /* tslint:enable */
+    result = render.fillInSpec(spec, definition)
+  })
+  test('it should set category field', () => {
+    expect(result.categoryField).toEqual('categoryField')
+  })
+  test('it should set graph stack and value field properties', () => {
+    expect(result.graphs.length).toEqual(3)
+    result.graphs.forEach((graph, i) => {
+      expect(graph.newStack).toBe(false)
+      expect(graph.valueField).toEqual(`Number_of_SUM_${i}`)
+    })
+  })
+})
+
+describe('when filling in a pie spec', () => {
+  let result
+  beforeAll(() => {
+    const spec = (pie as any)
+    /* tslint:disable */
+    const definition = {
+      "type": "pie",
+      "datasets": [
+        {
+          "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
+          "query": {
+            "orderByFields": "Number_of_SUM DESC",
+            "groupByFieldsForStatistics": "Type",
+            "outStatistics": [
+              {
+                "statisticType": "sum",
+                "onStatisticField": "Number_of",
+                "outStatisticFieldName": "Number_of_SUM"
+              }
+            ]
+          }
+        }
+      ],
+      "series": [
+        {
+          "category": {"field": "Type", "label": "Type"},
+          "value": {
+            "field": "Number_of_SUM",
+            "label": "Number of Students"
+          }
+        }
+      ]
+    }
+    /* tslint:enable */
+    result = render.fillInSpec(spec, definition)
+  })
+  test('it should set title field', () => {
+    expect(result.titleField).toEqual('Type')
+  })
+  test('it should set title field', () => {
+    expect(result.valueField).toEqual('Number_of_SUM')
+  })
+})
+
+describe('when passed an interim (v0.9.x) bar definition', () => {
+  let result
+  beforeAll(() => {
+    const spec = (bar as any)
+    /* tslint:disable */
+    const definition = {
+      "type": "bar", "datasets": [{
+        "url": "https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0",
+        "query": {
+          "groupByFieldsForStatistics": "Type", "orderByFields":
+            "Number_of_SUM DESC", "outStatistics": [{
+              "statisticType": "sum",
+              "onStatisticField": "Number_of", "outStatisticFieldName": "Number_of_SUM"
+            }]
+        }
+      }], "series": [{
+        "category": { "field": "Type", "label": "Facility Use" },
+        "value": { "field": "Number_of_SUM", "label": "Total Students" }
+      }]
+    }
+    /* tslint:enable */
+    result = render.fillInSpec(spec, definition)
+  })
+  test('should use series category field ', () => {
+    expect(result.categoryField).toEqual('Type')
+  })
+  test('should disable legend', () => {
+    expect(result.legend.enabled).toEqual(false)
+  })
+  test('should set x axis', () => {
+    expect(result.categoryAxis.title).toEqual('Facility Use')
+  })
+  test('should set y axis', () => {
+    const yAxis = result.valueAxes[0]
+    expect(yAxis.title).toEqual('Total Students')
+    expect(yAxis.position).toEqual('left')
+  })
+})
+
+describe('when passed an interim (v0.9.x) scatter definition', () => {
+  let result
+  beforeAll(() => {
+    const spec = (scatter as any)
+    /* tslint:disable */
+    const definition = {
+      "type":"scatter",
+      "datasets": [
+        {
+          "url":"https://services.arcgis.com/uDTUpUPbk8X8mXwl/arcgis/rest/services/Public_Schools_in_Onondaga_County/FeatureServer/0"
+        }
+      ],
+      "series": [
+        {
+          "category": { "field": "Number_of", "label": "Student Enrollment (2008)" },
+          "value": { "field":"F_of_teach", "label": "Fraction of Teachers" },
+          "color": { "field":"Type", "label":"Facility Type" }
+        }
+      ]
+    }
+    /* tslint:enable */
+    result = render.fillInSpec(spec, definition)
+  })
+  test('should use series category field ', () => {
+    expect(result.categoryField).toEqual('Number_of')
+  })
+  test('should disable legend', () => {
+    expect(result.legend.enabled).toEqual(false)
+  })
+  test('should set x axis', () => {
+    const xAxis = result.valueAxes[0]
+    expect(xAxis.position).toEqual('bottom')
+    expect(xAxis.title).toEqual('Student Enrollment (2008)')
+  })
+  test('should set y axis', () => {
+    const yAxis = result.valueAxes[1]
+    expect(yAxis.position).toEqual('left')
+    expect(yAxis.title).toEqual('Fraction of Teachers')
+  })
+})
+
+
 
 // NOTE: renderChart calls dependencies like amcharts and deepmerge
 // this suite used to run and presumeably actually rendered a chart somewhere
@@ -59,8 +313,8 @@ xdescribe('Rnder properly renders charts...', () => {
       ],
       series: [
         {
-          category: {field: 'Type', label: 'Type'},
-          value: {field: 'Number_of_SUM', label: 'Number of Students'}
+          category: { field: 'Type', label: 'Type' },
+          value: { field: 'Number_of_SUM', label: 'Number of Students' }
         }
       ],
       overrides: {