+++ packages/cedar/src/Chart.ts
@@ -1,5 +1,5 @@
 import { cedarAmCharts } from '@esri/cedar-amcharts'
-import { flattenFeatures } from './flatten/flatten'
+import { flattenFeatureSet, flattenFeatureSets } from './flatten/flatten'
 import { getData } from './query/query'
 import { createQueryParams, getQueryUrl } from './query/url'
 
@@ -7,29 +7,34 @@ function clone(json) {
   return JSON.parse(JSON.stringify(json))
 }
 
-// TODO: move this to flatten, or?
-function getChartData(datasets: IDataset[], series: ISeries[], datasetsData?: {}) {
+// flatten data from all datasets into a single table
+function getChartData(datasets: IDataset[], datasetsData?: {}) {
+  let name
+  let featureSet
+  if (datasets.length === 1) {
+    // if there's only one dataset, just return the flattened features
+    const dataset = datasets[0]
+    name = dataset.name || `dataset0`
+    featureSet = dataset.data || datasetsData[name]
+    return flattenFeatureSet(featureSet)
+  }
+
+  // if more than one, need to join the feature sets
   const featureSets = []
   const joinKeys = []
 
   // get array of featureSets from datasets data or datasetsData
   datasets.forEach((dataset, i) => {
     // TODO: make name required on datasets, or required if > 1 dataset?
-    const name = dataset.name || `dataset${i}`
+    name = dataset.name || `dataset${i}`
     // if dataset doesn't have inline data use data that was passed in
-    const featureSet = dataset.data || datasetsData[name]
+    featureSet = dataset.data || datasetsData[name]
     if (featureSet) {
       featureSets.push(featureSet)
     }
-    // TODO: this should not assume a 1:1 relationship between datasets and series
-    joinKeys.push(series[i].category.field)
+    joinKeys.push(dataset.join)
   })
-
-  // flatten data from all datasets into a single table
-  // if there's only one dataset, just flatten the features
-  // if more than one, need to join or append the feature sets
-  // return (featureSets.length === 1) ? flatten(featureSets[0]) : (joinKeys.length > 0) ? join(featureSets, joinKeys) : append(featureSets)
-  return flattenFeatures(featureSets, joinKeys)
+  return flattenFeatureSets(featureSets, joinKeys)
 }
 
 // TODO: where should these interfaces live?
@@ -38,7 +43,7 @@ export interface IDataset {
   url?: string,
   data?: any[],
   query: {},
-  append?: boolean
+  join?: string
 }
 
 export interface IField {
@@ -172,7 +177,7 @@ export default class Chart {
   // update chart from inline data and query responses
   public updateData(datasetsData) {
     const datasets = this.datasets()
-    this._data = datasets ? getChartData(datasets, this.series(), datasetsData) : []
+    this._data = datasets ? getChartData(datasets, datasetsData) : []
     return this
   }